requirements.txt
opencv-python==4.11.0.86
streamlit==1.29.1
streamlit-drawable-canvas==0.10.0
numpy==1.26.2
Pillow==11.0.0
replicate==0.1.0
google-genai==1.7.0